refactor(x-about-section): rename to XAboutCard
