+++ packages/kuma-gui/src/app/services/views/MeshServiceDetailView.vue
@@ -16,7 +16,7 @@
   >
     <AppView>
       <XLayout type="stack">
-        <XAboutSection
+        <XAboutCard
           :title="t('services.mesh-service.about.title')"
           :created="props.data.creationTime"
           :modified="props.data.modificationTime"
@@ -121,7 +121,7 @@
               </template>
             </template>
           </DefinitionCard>
-        </XAboutSection>
+        </XAboutCard>
 
         <div>
           <h3>
