+++ packages/kuma-gui/src/app/x/components/x-about-card/XAboutCard.vue
